refactor: separated highlight area to make it obvious they aren't really dependant - PLG-226
