+++ source/javascripts/components/ProductTour/Highlighter.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { useCallback, useEffect, useRef, useState } from "react";
+
 import { Portal } from "./Portal";
 
 import "./Highlighter.scss";
@@ -11,62 +11,6 @@ interface HighlighterProps {
 	isOpen: boolean;
 	renderSelection?: () => JSX.Element;
 }
-
-export const useHighlighter = (selectedId: string | undefined) => {
-	const ref = useRef<HTMLElement | null>(null);
-	const [highlightedRect, setHighlightedRect] = useState<DOMRect | null>(null);
-
-	const highlightElement = useCallback(id => {
-		const element = document.getElementById(id);
-		if (element) {
-			ref.current = element;
-			const rect = element.getBoundingClientRect();
-			setHighlightedRect(rect);
-		}
-	}, []);
-
-	useEffect(() => {
-		if (selectedId) {
-			highlightElement(selectedId);
-		}
-	}, [selectedId, highlightElement]);
-
-	useEffect(() => {
-		const onResize = () => {
-			if (ref.current) {
-				const rect = ref.current!.getBoundingClientRect();
-				setHighlightedRect(rect);
-			}
-		};
-
-		window.addEventListener("resize", onResize);
-
-		return () => {
-			window.removeEventListener("resize", onResize);
-		};
-	}, []);
-
-	let clipPath = "";
-	if (highlightedRect) {
-		const { x, y, width, height } = highlightedRect as DOMRect;
-		clipPath = `polygon(0% 0%,
-      0% 100%,
-      ${x}px 100%,
-      ${x}px ${y}px,
-      ${x + width}px ${y}px,
-      ${x + width}px ${y + height}px,
-      ${x}px ${y + height}px,
-      ${x}px 100%,
-      100% 100%,
-      100% 0%)`;
-	}
-
-	return {
-		clipPath,
-		rect: highlightedRect,
-		highlighted: !!clipPath
-	};
-};
 
 export const Highlighter = ({ clipPath, rect, children, isOpen, renderSelection }: HighlighterProps) => {
 	if (!isOpen || !clipPath) {
